requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.6
python-multipart==0.0.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.100.0
python-dotenv==1.0.0
itsdangerous==2.2.0
pydantic==2.11.7 
psycopg[binary]==3.2.9
sqlalchemy==2.0.35
pytz==2023.3

# テスト用パッケージ
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# オプション（パフォーマンス向上）
# pytest-xdist==3.3.1  # 並列実行
# pytest-cov==4.1.0    # カバレッジレポート 